requirements.txt
Django>=3.2,<4.0
djangorestframework
psycopg2-binary
uvicorn>=0.11.3,<1.0.0
